api/requirements.txt
beautifulsoup4==4.11.1
Django==3.2.1
django_environ==0.10.0
django_rest_knox==4.2.0
djangorestframework==3.12.4
djangorestframework_simplejwt==5.2.2
djongo==1.3.6
pymongo==3.12.3
# pymongo==4.3.3
Requests==2.30.0
